Typo(#123): fetch 모듈 함수 이름 변경 및 적용
